Initial commit
I believe this is ready for release so I am going to put it on Github.
In "plots" folder are pdf and png hurricane season timeline plots for
each year from 1851 through 2017, and a color legend.
In "data" folder is the NOAA HURDAT2 Atlantic database, and the same
data split up into individual files for each storm in directories for
each year.
In top directory are iPython notebooks to split up the dataset into
individual files, make and save the plots, and define retired storm
names (these are bold in the plots).